Allow configuring Forwarded/X-Forwarded-* headers
Implement the ROUTER_SET_FORWARDED_HEADERS environment variable and
haproxy.router.openshift.io/set-forwarded-headers route annotation.  The
annotation takes precedence over the environment variable.  The environment
variable and annotation may specify any of the following values to control
when HAProxy sets the Forwarded and X-Forwarded-* headers:

* "append" to append to existing headers.
* "replace" to replace existing headers.
* "if-none" to set the headers if they are absent.
* "never" to never set headers.

The default setting is "append".

* images/router/haproxy/conf/haproxy-config.template: Add a new variable,
setForwardedHeadersPattern, with a regular expression that matches the
values listed above.  Add a new variable, setForwardedHeadersAnnotation,
with the new annotation.  Add a new variable,
setForwardedHeadersDefaultValue, with value of ROUTER_SET_FORWARDED_HEADERS
if specified or else the default value "append".  Use these variables to
control when the Forwarded and X-Forwarded-* headers are set for the
backend.